Slit up S3-provider and S3-Share-provider
[#172225454]
https://www.pivotaltracker.com/story/show/172225454
