+++ package.json
@@ -30,6 +30,7 @@
     "gift": "0.10.2"
   },
   "devDependencies": {
+    "@types/aws-sdk": "^2.7.0",
     "@types/jest": "^25.2.1",
     "@types/pako": "^1.0.1",
     "@types/react": "^16.9.34",
